Менеджер вынесен в отдельный файл
